backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
requests==2.32.3
chromadb==0.6.3
langchain==0.3.25
langchain-community==0.3.23
sentence-transformers==3.4.1
python-dotenv==1.0.1
datasets
google-genai
